Silenced some warnings

+++ systemmenu.cc
@@ -1,3 +1,20 @@
+/*
+ * Mediabox: a light mediacenter solution
+ * Copyright (C) 2009 Jakob Truelsen
+ *
+ * This program is free software: you can redistribute it and/or modify
+ * it under the terms of the GNU General Public License as published by
+ * the Free Software Foundation, either version 3 of the License, or
+ * (at your option) any later version.
+ *
+ * This program is distributed in the hope that it will be useful,
+ * but WITHOUT ANY WARRANTY; without even the implied warranty of
+ * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the
+ * GNU General Public License for more details.
+ *
+ * You should have received a copy of the GNU General Public License
+ * along with this program.  If not, see <http://www.gnu.org/licenses/>
+ */
 #include "systemmenu.hh"
 #include <vector>
 #include "config.hh"
@@ -177,8 +194,8 @@ public:
 			Color(0,0,150,210) );
 		f2->setRadius(0.02);
 		
-		int x;
-		int y;
+		size_t x;
+		size_t y;
 		for(x=2; ; ++x) {
 			for(y=x-1; y <= x; ++y)
 				if(x*y >= parts.size() + 1) goto good;
@@ -188,8 +205,8 @@ public:
 		double dy = (0.8-0.3)/x;
 		parts_pitch = x;
 		
-		for(int j=0; j < y; ++j) 
-			for(int i=0; i < x; ++i) 
+		for(size_t j=0; j < y; ++j) 
+			for(size_t i=0; i < x; ++i) 
 				l.push_back( make_pair(i*dx+0.25, j*dy+0.25) );
 				
 		for(size_t i=0; i < parts.size(); ++i) {
